chore(test): extract test util into shared files
