+++ packages/jit/test/unit/util.ts
@@ -1,0 +1,3 @@
+import { _, stringify, jsonStringify, htmlStringify, createElement, verifyEqual } from '../../../../scripts/test-lib';
+
+export { _, stringify, jsonStringify, htmlStringify, createElement, verifyEqual };
